--- a/clovers/src/camera.rs
+++ b/clovers/src/camera.rs
@@ -3,7 +3,8 @@
 #![allow(clippy::too_many_arguments)] // TODO: Camera::new() has a lot of arguments.
 
 use crate::{random::random_in_unit_disk, ray::Ray, Float, Vec3, PI};
-use rand::prelude::*;
+use rand::rngs::SmallRng;
+use rand::{Rng, SeedableRng};
 use serde::{Deserialize, Serialize};
 
 #[derive(Copy, Clone, Debug, Serialize, Deserialize)]
@@ -96,7 +97,7 @@ impl Camera {
 
     // TODO: fix the mysterious (u,v) vs (s,t) change that came from the tutorial
     /// Generates a new [Ray] from the camera, at a random location of the aperture, at a random time interval between `time_0`, `time_1` of the camera.
-    pub fn get_ray(self, s: Float, t: Float, rng: &mut ThreadRng) -> Ray {
+    pub fn get_ray(self, s: Float, t: Float, rng: &mut SmallRng) -> Ray {
         // TODO: add a better defocus blur / depth of field implementation
         let rd: Vec3 = self.lens_radius * random_in_unit_disk(rng);
         let offset: Vec3 = self.u * rd.x + self.v * rd.y;