--- a/clovers/src/colorize.rs
+++ b/clovers/src/colorize.rs
@@ -8,16 +8,11 @@ use crate::{
     scenes::Scene,
     Float, EPSILON_SHADOW_ACNE,
 };
-use rand::prelude::*;
+use rand::rngs::SmallRng;
+use rand::{Rng, SeedableRng};
 
 /// The main coloring function. Sends a [Ray] to the [Scene], sees if it hits anything, and eventually returns a [Color]. Taking into account the [Material](crate::materials::Material) that is hit, the method recurses with various adjustments, with a new [Ray] started from the location that was hit.
-pub fn colorize(
-    ray: &Ray,
-    scene: &Scene,
-    depth: u32,
-    max_depth: u32,
-    rng: &mut ThreadRng,
-) -> Color {
+pub fn colorize(ray: &Ray, scene: &Scene, depth: u32, max_depth: u32, rng: &mut SmallRng) -> Color {
     // Have we reached the maximum recursion i.e. ray bounce depth?
     if depth > max_depth {
         // Ray bounce limit reached, early return background_color