--- a/clovers/src/objects/flip_face.rs
+++ b/clovers/src/objects/flip_face.rs
@@ -6,7 +6,8 @@ use crate::{
     ray::Ray,
     Box, Float,
 };
-use rand::prelude::*;
+use rand::rngs::SmallRng;
+use rand::{Rng, SeedableRng};
 use serde::{Deserialize, Serialize};
 
 use super::Object;
@@ -40,7 +41,7 @@ impl FlipFace {
         ray: &Ray,
         distance_min: Float,
         distance_max: Float,
-        rng: &mut ThreadRng,
+        rng: &mut SmallRng,
     ) -> Option<HitRecord> {
         self.object
             .hit(ray, distance_min, distance_max, rng)