--- a/clovers/src/objects/sphere.rs
+++ b/clovers/src/objects/sphere.rs
@@ -4,7 +4,8 @@ use crate::{
     aabb::AABB, hitable::HitRecord, materials::Material, onb::ONB, random::random_to_sphere,
     ray::Ray, Float, Vec3, EPSILON_SHADOW_ACNE, PI,
 };
-use rand::prelude::*;
+use rand::rngs::SmallRng;
+use rand::{Rng, SeedableRng};
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Copy, Debug, Deserialize, Serialize)]
@@ -53,7 +54,7 @@ impl Sphere {
         ray: &Ray,
         distance_min: Float,
         distance_max: Float,
-        _rng: &mut ThreadRng,
+        _rng: &mut SmallRng,
     ) -> Option<HitRecord> {
         let oc: Vec3 = ray.origin - self.center;
         let a: Float = ray.direction.norm_squared();
@@ -113,7 +114,7 @@ impl Sphere {
     }
 
     /// Returns the probability density function for the sphere? TODO: what does this do again and how
-    pub fn pdf_value(&self, origin: Vec3, vector: Vec3, time: Float, rng: &mut ThreadRng) -> Float {
+    pub fn pdf_value(&self, origin: Vec3, vector: Vec3, time: Float, rng: &mut SmallRng) -> Float {
         match self.hit(
             &Ray::new(origin, vector, time),
             EPSILON_SHADOW_ACNE,
@@ -134,7 +135,7 @@ impl Sphere {
 
     // TODO: understand, document
     /// Utility function from Ray Tracing: The Rest of Your Life. TODO: understand, document
-    pub fn random(&self, origin: Vec3, rng: &mut ThreadRng) -> Vec3 {
+    pub fn random(&self, origin: Vec3, rng: &mut SmallRng) -> Vec3 {
         let direction: Vec3 = self.center - origin;
         let distance_squared: Float = direction.norm_squared();
         let uvw = ONB::build_from_w(direction);