--- a/clovers/src/pdf.rs
+++ b/clovers/src/pdf.rs
@@ -3,7 +3,8 @@
 #![allow(missing_docs)] // TODO: Lots of undocumented things for now
 
 use crate::{hitable::Hitable, onb::ONB, random::random_cosine_direction, Box, Float, Vec3, PI};
-use rand::prelude::*;
+use rand::rngs::SmallRng;
+use rand::{Rng, SeedableRng};
 
 #[derive(Debug)]
 pub enum PDF<'a> {
@@ -14,7 +15,7 @@ pub enum PDF<'a> {
 }
 
 impl<'a> PDF<'a> {
-    pub fn value(&self, direction: Vec3, time: Float, rng: &mut ThreadRng) -> Float {
+    pub fn value(&self, direction: Vec3, time: Float, rng: &mut SmallRng) -> Float {
         match self {
             PDF::CosinePDF(p) => p.value(direction, time, rng),
             PDF::HitablePDF(p) => p.value(direction, time, rng),
@@ -22,7 +23,7 @@ impl<'a> PDF<'a> {
             PDF::ZeroPDF(p) => p.value(direction, time, rng),
         }
     }
-    pub fn generate(&self, rng: &mut ThreadRng) -> Vec3 {
+    pub fn generate(&self, rng: &mut SmallRng) -> Vec3 {
         match self {
             PDF::CosinePDF(p) => p.generate(rng),
             PDF::HitablePDF(p) => p.generate(rng),
@@ -44,7 +45,7 @@ impl CosinePDF {
         }
     }
 
-    pub fn value(&self, direction: Vec3, _time: Float, _rng: &mut ThreadRng) -> Float {
+    pub fn value(&self, direction: Vec3, _time: Float, _rng: &mut SmallRng) -> Float {
         let cosine = direction.normalize().dot(&self.uvw.w);
         if cosine <= 0.0 {
             0.0
@@ -53,7 +54,7 @@ impl CosinePDF {
         }
     }
 
-    pub fn generate(&self, rng: &mut ThreadRng) -> Vec3 {
+    pub fn generate(&self, rng: &mut SmallRng) -> Vec3 {
         self.uvw.local(random_cosine_direction(rng))
     }
 }
@@ -69,11 +70,11 @@ impl<'a> HitablePDF<'a> {
         HitablePDF { origin, hitable }
     }
 
-    pub fn value(&self, direction: Vec3, time: Float, rng: &mut ThreadRng) -> Float {
+    pub fn value(&self, direction: Vec3, time: Float, rng: &mut SmallRng) -> Float {
         self.hitable.pdf_value(self.origin, direction, time, rng)
     }
 
-    pub fn generate(&self, rng: &mut ThreadRng) -> Vec3 {
+    pub fn generate(&self, rng: &mut SmallRng) -> Vec3 {
         self.hitable.random(self.origin, rng)
     }
 }
@@ -93,11 +94,11 @@ impl<'a> MixturePDF<'a> {
         }
     }
 
-    pub fn value(&self, direction: Vec3, time: Float, rng: &mut ThreadRng) -> Float {
+    pub fn value(&self, direction: Vec3, time: Float, rng: &mut SmallRng) -> Float {
         0.5 * self.pdf1.value(direction, time, rng) + 0.5 * self.pdf2.value(direction, time, rng)
     }
 
-    pub fn generate(&self, rng: &mut ThreadRng) -> Vec3 {
+    pub fn generate(&self, rng: &mut SmallRng) -> Vec3 {
         if rng.gen::<bool>() {
             self.pdf1.generate(rng)
         } else {
@@ -115,11 +116,11 @@ impl ZeroPDF {
         ZeroPDF {}
     }
 
-    pub fn value(&self, _direction: Vec3, _time: Float, _rng: &mut ThreadRng) -> Float {
+    pub fn value(&self, _direction: Vec3, _time: Float, _rng: &mut SmallRng) -> Float {
         0.0
     }
 
-    pub fn generate(&self, _rng: &mut ThreadRng) -> Vec3 {
+    pub fn generate(&self, _rng: &mut SmallRng) -> Vec3 {
         Vec3::new(1.0, 0.0, 0.0)
     }
 }