--- a/clovers/src/random.rs
+++ b/clovers/src/random.rs
@@ -1,10 +1,11 @@
 //! Various internal helper functions for getting specific kinds of random values.
 
 use crate::{Float, Vec3, PI};
-use rand::prelude::*;
+use rand::rngs::SmallRng;
+use rand::{Rng, SeedableRng};
 
 /// Internal helper. Originally used for lambertian reflection with flaws
-pub fn random_in_unit_sphere(rng: &mut ThreadRng) -> Vec3 {
+pub fn random_in_unit_sphere(rng: &mut SmallRng) -> Vec3 {
     let mut position: Vec3;
     // TODO: figure out a non-loop method
     // See https://github.com/RayTracing/raytracing.github.io/issues/765
@@ -17,7 +18,7 @@ pub fn random_in_unit_sphere(rng: &mut ThreadRng) -> Vec3 {
 }
 
 /// Internal helper. Use this for the more correct "True Lambertian" reflection
-pub fn random_unit_vector(rng: &mut ThreadRng) -> Vec3 {
+pub fn random_unit_vector(rng: &mut SmallRng) -> Vec3 {
     let a: Float = rng.gen_range(0.0..2.0 * PI);
     let z: Float = rng.gen_range(-1.0..1.0);
     let r: Float = (1.0 - z * z).sqrt();
@@ -25,7 +26,7 @@ pub fn random_unit_vector(rng: &mut ThreadRng) -> Vec3 {
 }
 
 /// Internal helper.
-pub fn random_in_unit_disk(rng: &mut ThreadRng) -> Vec3 {
+pub fn random_in_unit_disk(rng: &mut SmallRng) -> Vec3 {
     let mut position: Vec3;
     // TODO: figure out a non-loop method
     // See https://github.com/RayTracing/raytracing.github.io/issues/765
@@ -39,7 +40,7 @@ pub fn random_in_unit_disk(rng: &mut ThreadRng) -> Vec3 {
 }
 
 /// Internal helper.
-pub fn random_cosine_direction(rng: &mut ThreadRng) -> Vec3 {
+pub fn random_cosine_direction(rng: &mut SmallRng) -> Vec3 {
     let r1 = rng.gen::<Float>();
     let r2 = rng.gen::<Float>();
     let z = (1.0 - r2).sqrt();
@@ -52,7 +53,7 @@ pub fn random_cosine_direction(rng: &mut ThreadRng) -> Vec3 {
 }
 
 /// Internal helper.
-pub fn random_to_sphere(radius: Float, distance_squared: Float, rng: &mut ThreadRng) -> Vec3 {
+pub fn random_to_sphere(radius: Float, distance_squared: Float, rng: &mut SmallRng) -> Vec3 {
     let r1 = rng.gen::<Float>();
     let r2 = rng.gen::<Float>();
     let z = 1.0 + r2 * ((1.0 - radius * radius / distance_squared).sqrt() - 1.0);