- remove some unused dependencies and features
- update indicatif
- use smallrng instead of threadrng

perf testing with `hyperfine "just scene"`
```
this branch
Benchmark #1: just scene
  Time (mean ± σ):      6.952 s ±  0.601 s    [User: 422.620 s, System: 3.744 s]
  Range (min … max):    6.237 s …  7.809 s    10 runs
 
hyperfine "just scene"  4226.38s user 37.49s system 6115% cpu 1:09.72 total


main
Benchmark #1: just scene
  Time (mean ± σ):      8.359 s ±  0.146 s    [User: 514.850 s, System: 0.616 s]
  Range (min … max):    8.193 s …  8.514 s    10 runs
 
hyperfine "just scene"  5148.68s user 6.22s system 6152% cpu 1:23.79 total
```